pubsys: refactor ssm module
This refactor is needed to let the `ssm` pubsys command append to the
existing SSM parameters metadata file using the existing code.

Before this change, it wasn't possible to re-use the same functions used
to append SSM parameters to existing files due to circular dependencies
on errors and modules. With this refactor, the functions were moved to
the `ssm` module to allow other submodules use this functionality.

Signed-off-by: Arnaldo Garcia Rincon <agarrcia@amazon.com>